requirements.txt
attrs==19.3.0
certifi==2020.6.20
coverage==5.2.1
Django==2.2.14
djangorestframework==3.11.0
Faker==0.9.1
importlib-metadata==1.7.0
mixer==6.1.3
more-itertools==8.4.0
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.2
wcwidth==0.2.5
zipp==3.1.0
